beginning to clean some of the data
a lot of the data is too large to be in the repository, but will transfer what I can. Hopefully I can make cleaned summary tables that can got in the repository. Would perfer the code to be in one place instead of two... Maybe I'll figure out away to keep the data on my local device and then the code in the cloned folder.... we will see